+++ drivers/net/ethernet/mscc/ocelot_net.c
@@ -801,7 +801,7 @@ nla_put_failure:
 static int ocelot_port_fdb_dump(struct sk_buff *skb,
 				struct netlink_callback *cb,
 				struct net_device *dev,
-				struct net_device *filter_dev, int *idx)
+				struct net_device *filter_dev, u64 *idx)
 {
 	struct ocelot_port_private *priv = netdev_priv(dev);
 	struct ocelot *ocelot = priv->port.ocelot;
